backend/requirements.txt
Django>=5.0
gunicorn
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg[binary]
redis==5.0.1
django-redis==5.4.0
django-cors-headers==4.3.1
python-decouple==3.8
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.1.0
dj-database-url-3.0.1
setuptools>=70.0.0
